Option to load multiple language files
With this commit, multiple language files can be set for loading. If
only a string is passed to langFile, then only this one language file is
loaded. However, if an array is passed, multiple language files will be
loaded. Each language file must be as a string in this array.